refactor(export): extract frame compositing block from renderFrame
Moves the ~385-line z-order compositing block (WebGPU blend-mode compositor
path + Canvas2D fallback, with pooled GPU-texture/canvas lifetime management)
into compositeFrameResults in frame-compositing.ts. renderFrame now calls it and
uses the returned final-composite source. Body is verbatim; the five per-task
helpers are passed in via FrameCompositingDeps. Re-exports FrameRenderTask
through the composition-runtime deps adapter to respect feature boundaries.

renderFrame: 627 -> 453 lines; client-render-engine.ts: 2229 -> 2053.

VERIFICATION REQUIRED: the WebGPU compositor path is not covered by the jsdom
test suite. Verify export + preview output visually (blend modes, masks,
transitions, occlusion) before merging.
