+++ src/features/export/utils/client-render-engine.ts
@@ -33,8 +33,6 @@ import { VideoSourcePool } from '@/features/export/deps/player-contract'
 // Import subsystems
 import { buildKeyframesMap } from './canvas-keyframes'
 import { type AdjustmentLayerWithTrackOrder } from './canvas-effects'
-import { DEFAULT_LAYER_PARAMS } from '@/infrastructure/gpu-compositor'
-import type { CompositeLayer } from '@/infrastructure/gpu-compositor'
 import { GpuPipelineManager } from './gpu-pipeline-manager'
 import { isItemFullyOccluding, type FrameOcclusionContext } from './frame-occlusion'
 import {
@@ -47,8 +45,8 @@ import {
   renderItemWithEffects as renderItemWithEffectsPure,
   type FrameItemRenderDeps,
 } from './frame-render-tasks'
+import { compositeFrameResults } from './frame-compositing'
 import {
-  applyMasks,
   buildMaskFrameIndex,
   getActiveMasksForFrame,
   type MaskCanvasSettings,
@@ -1641,200 +1639,26 @@ export async function createCompositionRenderer(
           await itemRenderContext.gpuPipeline.waitForSubmittedWork()
         }
 
-        // Composite all results in z-order (preserved by renderTasks ordering)
-        if (useGpuCompositor && gpu.compositor && gpu.maskManager && gpuCompositeOutput) {
-          // GPU compositing path — pixel-perfect blend modes via WebGPU
-          const device = gpu.effects!.getDevice()
-          const w = canvasSettings.width
-          const h = canvasSettings.height
-          const layers: CompositeLayer[] = []
-          const layerTextures: GPUTexture[] = []
-          const layerMaskTextures: GPUTexture[] = []
-          const compositedResults: Array<{
-            task: (typeof renderTasks)[number]
-            result: RenderedTaskResult
-            fallbackMasks: typeof activeMasks
-          }> = []
-
-          for (let i = 0; i < results.length; i++) {
-            const task = renderTasks[i]!
-            const taskHasCornerPin =
-              task.type === 'item'
-                ? hasCornerPin(getCurrentItem(task.item).cornerPin)
-                : hasCornerPin(getCurrentItem(task.transition.leftClip).cornerPin) ||
-                  hasCornerPin(getCurrentItem(task.transition.rightClip).cornerPin)
-            const applicableMasks = activeMasks.filter((mask) =>
-              doesMaskAffectTrack(mask.trackOrder, task.trackOrder),
-            )
-            const shouldUseSeparateMask = task.type === 'item' && !taskHasCornerPin
-            let result: RenderedTaskResult | null = shouldUseSeparateMask
-              ? (results[i] ?? null)
-              : applyTrackScopedMasks(results[i] ?? null, task.trackOrder, taskHasCornerPin)
-            if (!result) continue
-
-            let maskInfo =
-              shouldUseSeparateMask && applicableMasks.length > 0
-                ? renderMasksToGpuTexture(applicableMasks)
-                : null
-            let fallbackMasks = shouldUseSeparateMask ? applicableMasks : []
-            if (shouldUseSeparateMask && applicableMasks.length > 0 && !maskInfo) {
-              const maskedResult = applyTrackScopedMasks(result, task.trackOrder, false)
-              if (!maskedResult) continue
-              result = maskedResult
-              fallbackMasks = []
-            }
-
-            const blendMode =
-              task.type === 'item'
-                ? (getEffectiveBlendMode(getCurrentItem(task.item)) ?? 'normal')
-                : 'normal'
-
-            // Upload item canvas to GPU texture (pooled — no per-frame alloc)
-            let tex = result.gpuTexture
-            if (!tex) {
-              if (!result.source) continue
-              tex = gpu.texturePool!.acquire(w, h)
-              device.queue.copyExternalImageToTexture(
-                { source: result.source, flipY: false },
-                { texture: tex },
-                { width: w, height: h },
-              )
-            }
-            layerTextures.push(tex)
-
-            if (maskInfo) {
-              layerMaskTextures.push(maskInfo.texture)
-            }
-
-            compositedResults.push({
-              task,
-              result,
-              fallbackMasks,
-            })
-
-            layers.push({
-              params: {
-                ...DEFAULT_LAYER_PARAMS,
-                blendMode,
-                sourceAspect: w / h,
-                outputAspect: w / h,
-                hasMask: Boolean(maskInfo),
-              },
-              textureView: tex.createView(),
-              maskView: maskInfo?.view ?? gpu.maskManager.getFallbackView(),
-            })
-          }
-
-          let compositedToGpuCanvas = false
-          if (layers.length > 0) {
-            try {
-              compositedToGpuCanvas = gpu.compositor.compositeToCanvas(
-                layers,
-                w,
-                h,
-                gpuCompositeOutput.ctx,
-              )
-            } catch (error) {
-              getLog().warn('GPU compositor failed - using Canvas2D blend fallback', {
-                error,
-              })
-            }
-          }
-
-          if (compositedToGpuCanvas) {
-            await itemRenderContext.gpuPipeline?.waitForSubmittedWork()
-            finalCompositeSource = gpuCompositeOutput.canvas
-          } else {
-            // Fall back to the established Canvas2D compositor if the GPU target
-            // isn't available for this frame. This preserves feature parity and
-            // avoids dropping content when WebGPU canvas presentation fails.
-            for (const { task, result, fallbackMasks } of compositedResults) {
-              let fallbackResult = result
-              if (!fallbackResult.source && task.type === 'transition') {
-                fallbackResult = await renderTransitionFallbackCanvas(task)
-              }
-              if (!fallbackResult.source && task.type === 'item') {
-                const rerenderedFallback = await renderItemWithEffects(
-                  task.item,
-                  task.trackOrder,
-                  true,
-                  contentCtx,
-                  true,
-                  false,
-                  false,
-                )
-                if (!rerenderedFallback) continue
-                fallbackResult = rerenderedFallback
-              }
-              let fallbackSource = fallbackResult.source
-              if (!fallbackSource) continue
-              if (fallbackMasks.length > 0) {
-                const { canvas: fallbackMaskedCanvas, ctx: fallbackMaskedCtx } =
-                  canvasPool.acquire()
-                applyMasks(fallbackMaskedCtx, fallbackSource, fallbackMasks, maskSettings)
-                fallbackResult = {
-                  source: fallbackMaskedCanvas,
-                  poolCanvases: [...fallbackResult.poolCanvases, fallbackMaskedCanvas],
-                }
-                fallbackSource = fallbackMaskedCanvas
-              }
-              const blendMode =
-                task.type === 'item' ? getEffectiveBlendMode(getCurrentItem(task.item)) : undefined
-              if (blendMode && blendMode !== 'normal') {
-                contentCtx.globalCompositeOperation = getCompositeOperation(blendMode)
-              }
-
-              contentCtx.drawImage(fallbackSource, 0, 0)
-
-              if (blendMode && blendMode !== 'normal') {
-                contentCtx.globalCompositeOperation = 'source-over'
-              }
-              if (fallbackResult !== result) {
-                for (const c of fallbackResult.poolCanvases) canvasPool.release(c)
-              }
-            }
-          }
-
-          for (const { result } of compositedResults) {
-            for (const c of result.poolCanvases) canvasPool.release(c)
-          }
-
-          // Release pooled textures (no GPU destroy — recycled next frame)
-          for (const tex of layerTextures) gpu.texturePool!.release(tex)
-          for (const tex of layerMaskTextures) gpu.texturePool!.release(tex)
-        } else {
-          // Canvas2D compositing fallback
-          for (let i = 0; i < results.length; i++) {
-            const task = renderTasks[i]!
-            const result = applyTrackScopedMasks(
-              results[i] ?? null,
-              task.trackOrder,
-              task.type === 'item'
-                ? hasCornerPin(getCurrentItem(task.item).cornerPin)
-                : hasCornerPin(getCurrentItem(task.transition.leftClip).cornerPin) ||
-                    hasCornerPin(getCurrentItem(task.transition.rightClip).cornerPin),
-            )
-            if (!result) continue
-            if (!result.source) {
-              if (result.gpuTexture) gpu.texturePool?.release(result.gpuTexture)
-              continue
-            }
-
-            const blendMode =
-              task.type === 'item' ? getEffectiveBlendMode(getCurrentItem(task.item)) : undefined
-            if (blendMode && blendMode !== 'normal') {
-              contentCtx.globalCompositeOperation = getCompositeOperation(blendMode)
-            }
-
-            contentCtx.drawImage(result.source, 0, 0)
-
-            if (blendMode && blendMode !== 'normal') {
-              contentCtx.globalCompositeOperation = 'source-over'
-            }
-
-            for (const c of result.poolCanvases) canvasPool.release(c)
-          }
-        }
+        finalCompositeSource = await compositeFrameResults({
+          useGpuCompositor,
+          gpu,
+          gpuCompositeOutput,
+          canvasSettings,
+          maskSettings,
+          renderTasks,
+          results,
+          activeMasks,
+          contentCanvas,
+          contentCtx,
+          itemRenderContext,
+          canvasPool,
+          getCurrentItem,
+          getEffectiveBlendMode,
+          applyTrackScopedMasks,
+          renderMasksToGpuTexture,
+          renderTransitionFallbackCanvas,
+          renderItemWithEffects,
+        })
       }
 
       // Log occlusion culling stats periodically (only in development)
